\name{geepack-internal}
\alias{internal}

\alias{tidy}
\alias{\%>\%}
\alias{crossutri}
\alias{genZodds}
\alias{anova.geeglm}
\alias{anovageePrim2}
\alias{anova.geeglmlist}
\alias{plot.geeglm}
\alias{print.geeglm}
\alias{eprint}
\alias{print.summary.geeglm}
\alias{residuals.geeglm}
\alias{summary.geeglm}
\alias{vcov.geeglm}


\title{Internal functions for the geepack package}
\description{Internal functions called by other functions.}

\usage{
crossutri(wave)
genZodds(clusz, waves, corstrv, ncat)
}
\details{
  These are not to be called directly by the user.
}

\keyword{internal}% at least one, from doc/KEYWORDS
