\name{geese-internal}
\alias{crossutri}
%\alias{genZcor}
\alias{genZodds}
\title{Internal geese functions}
\description{
  Internal functions called by other functions.
}
\usage{
crossutri(wave)
genZcor(clusz, waves, corstrv)
genZodds(clusz, waves, corstrv, ncat)
}
\details{
  These are not to be called directly by the user.
}
\keyword{internal}
