\name{SHDgeese-internal}
\alias{anova.geeglm}
\alias{anovageePrim2}
\alias{anova.geeglmlist}
\alias{plot.geeglm}
\alias{print.geeglm}
\alias{eprint}
\alias{print.summary.geeglm}
\alias{residuals.geeglm}
\alias{summary.geeglm}

\title{Internal geese functions}
\description{
  Internal functions called by other functions.
}
% \usage{
% crossutri(wave)
% genZcor(clusz, waves, corstrv)
% genZodds(clusz, waves, corstrv, ncat)
% }
% \details{
%   These are not to be called directly by the user.
% }
\keyword{internal}
