% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-geepack.R
\docType{data}
\name{sitka89}
\alias{sitka89}
\title{Growth of Sitka Spruce Trees}
\format{
A dataframe

\describe{
\item{size:}{size of the tree measured in \eqn{log(height*diamter^2)}}

\item{time:}{days after the 1st january, 1988}

\item{tree:}{id number of a tree}

\item{treat:}{ozone: grown under ozone environment, control: ozone free}

}
}
\usage{
sitka89
}
\description{
Impact of ozone on the growth of sitka spruce trees.
}
\examples{

data(sitka89)

}
\keyword{datasets}
