% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-geepack.R
\docType{data}
\name{koch}
\alias{koch}
\title{Ordinal Data from Koch}
\format{
This data frame contains the following columns:
\describe{
\item{trt}{a numeric vector}
\item{day}{a numeric vector}
\item{y}{an ordered factor with levels: \code{1} < \code{2} < \code{3}}
\item{id}{a numeric vector}
}
}
\usage{
koch
}
\description{
The \code{koch} data frame has 288 rows and 4 columns.
}
\examples{

data(koch)
fit <- ordgee(ordered(y) ~ trt + as.factor(day), id=id, data=koch, corstr="exch")
summary(fit)

}
\keyword{datasets}
